+++ HISTORY.md
@@ -1,4 +1,5 @@
 # History
-## 0.0.1 (2023-06-27)
-This is the first version.It provides the **VimeoConnection**, **Video** and **Album** classes.
-This package is still in beta, as many of the possible methods of this classes are not implemented yet.
+## 0.0.1 (2023-07-07)
+This is the first version.It provides the **VimeoConnection**, **VimeoData**, **Video** and **VimeoShowcase** and **VimeoFolder** classes.
+
+The package is still in beta, as many of the possible methods of this classes are not implemented yet and other classes (e.g. for Groups or Portfolios) shall be implemented later.